#1322 Change structure, handling moved to the job
